Rework property discovery for classes
Clearly split the handling of RLMObject subclasses from that of
RealmSwift.Object subclasses, rather than mixing the two in a way that made it
unclear which blocks of code were relevant for which, and as a result get rid
of the RLMObjectUtil methods only relevant to one of them. Pass all of the
relevant information to RLMProperty's initializer rather than having several
needlessly mutable properties.

Significantly redesign RealmSwift.Object property discovery. Get the list of
all property names from Swift reflection and correlate that with the obj-c
properties as needed rather than initially populating with the obj-c properties
and then tacking on several different kinds of Swift-only properties. This
simplifies the logic a bit by eliminating the "add a property, then remove it
and add a new optional version of that property" dance, and makes it so that
the declared property order is actually respected.